src/backend/requirements.txt
openai>=1.61.0
google-genai>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.10.0
colorama>=0.4.6
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
